requirements.txt
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.2.0
cabby==0.1.20
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
cloudmersive-convert-api-client==3.2.4
colorama==0.4.4
colorlog==4.1.0
configparser==5.0.1
crayons==0.4.0
cryptography==36.0.1
discord==1.7.3
discord.py==1.7.3
falcon==3.0.1
furl==2.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
gunicorn==20.1.0
hvac==0.10.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.6.0
inflect==4.1.0
jmespath==0.10.0
jsonschema==3.2.0
libmagic==1.0
libtaxii==1.1.117
lxml==4.5.1
multidict==5.2.0
ntplib==0.3.4
numpy==1.19.4
opencv-python==4.4.0.46
orderedmultidict==1.0.1
paho-mqtt==1.5.0
paramiko==2.9.2
parsedatetime==2.5
pexpect==4.8.0
ptyprocess==0.7.0
pyaes==1.6.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-magic==0.4.24
pytz==2019.3
redis==3.4.1
requests==2.25.0
schedule==1.1.0
selenium==3.141.0
simplejson==3.17.2
six==1.14.0
smmap==5.0.0
stdlib-list==0.6.0
stix2==2.0.2
stix2-patterns==1.3.1
taxii2-client==2.1.0
tcex==2.0.14
threatconnect==2.4.21
timeago==1.0.14
typing_extensions==4.0.1
tzlocal==2.0.0
urllib3==1.26.2
webdriver-manager==3.2.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.1.0
